feat : 로그인기능으로 게시물 CRUD 및 예외처리

+++ src/main/java/com/example/newsfeed/service/FeedService.java
@@ -7,17 +7,21 @@ import com.example.newsfeed.dto.FriendResponseDto;
 import com.example.newsfeed.dto.PagedFeedResponseDto;
 import com.example.newsfeed.entity.Feed;
 import com.example.newsfeed.entity.Friend;
+import com.example.newsfeed.entity.User;
 import com.example.newsfeed.repository.FeedRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -25,6 +29,7 @@ import java.util.stream.Collectors;
 public class FeedService {
     private final FeedRepository feedRepository;
     private final FriendService friendService;
+    private final UserService userService;
 
     public List<FeedResponseDto> allFeeds(){
         return feedRepository.findAll()
@@ -62,23 +67,30 @@ public class FeedService {
         return feedRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("잘못된 ID 값입니다."));
     }
 
-    public  void deleteFeed(Long id){
-        findFeedById(id);
+    public  void deleteFeed(Long id,Long userId){
+        Feed feed = findFeedById(id);
+        if(!Objects.equals(feed.getId(), userId)){
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "본인의 게시물만 삭제할 수 있습니다.");
+        }
         feedRepository.deleteById(id);
     }
 
     @Transactional
-    public FeedResponseDto createFeed(FeedRequestDto feedRequestDto){
-        Feed feed = new Feed(feedRequestDto.getNickName(),feedRequestDto.getUserId(),feedRequestDto.getPhone(),feedRequestDto.getEmail(),feedRequestDto.getContent());
+    public FeedResponseDto createFeed(Long userId, FeedRequestDto feedRequestDto){
+        User user = userService.findUserById(userId);
+        Feed feed = new Feed(userId,feedRequestDto.getContent(),user.getEmail(),user.getNickName(),user.getPhone());
         Feed saveFeed = feedRepository.save(feed);
         return FeedResponseDto.toDto(saveFeed);
     }
 
 
     @Transactional
-    public FeedResponseDto updateFeed(Long id, FeedRequestDto feedRequestDto){
+    public FeedResponseDto updateFeed(Long id, Long userId, FeedRequestDto feedRequestDto){
         Feed feed = findFeedById(id);
-        feed.update(feedRequestDto.getContent());
+        if(!Objects.equals(feed.getId(), userId)){
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "본인의 게시물만 수정할 수 있습니다.");
+        }
+        feed.setContent(feedRequestDto.getContent());
         return FeedResponseDto.toDto(feed);
     }
